Prevent "Expect: 100-continue" in Async HTTP POST
cURL default behaviour whan doing HTTP POST with a body larger than 1024 bytes is to use the "Expect: 100-continue" behaviour.
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/100
This cause a bug with the Async feature preventing the request body from being sent.
Adding "Expect:" to the header list prevent the default behaviour so Curl send the body without waiting for a "100-continue" response.